requirements.txt
jax
flax
distrax
jaxopt
numpy
absl-py
chex
ml_collections
clu
tensorflow
diffrax
